feat: 修改 NSVQQuantizer 中 codebooks_used 的数据类型为 float32，并在 VQAPTrainer 中调整分布式同步逻辑为平均值，优化代码可读性和性能

+++ model/module/nsvq.py
@@ -101,7 +101,7 @@ class NSVQQuantizer(nn.Module):
 		self.codebooks = nn.Parameter(torch.empty(self.codebook_size, self.codebook_dim))  # [K, D]
 		self.register_buffer(
 			"codebooks_used",
-			torch.zeros(self.codebook_size, dtype=torch.long),  # [K]，每个码字累计被分配的次数
+			torch.zeros(self.codebook_size, dtype=torch.float32),  # [K]，每个码字累计被分配的次数
 			persistent=True,
 		)
